requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
Flask-Cors==4.0.0
Flask-JSON==0.4.0
git+https://github.com/idoshr/flask-mongoengine.git@e244408acf440c4208f7ddcd6e5d819cb472e4da
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
openai==0.28.0
gunicorn==21.2.0
tensorflow==2.18.0
pillow==11.1.0
numpy==2.0.2